Please check this.
The actual calculation of "probability_of_infection" is exact as it is in your code, but I'm not sure about my logic behind the effectiveness.
I tried to mimic the effectiveness of a N95 mask. (Though slightly lower, at 90%, to include leakage because not everyone wears them correctly. Maybe it needs to be 80%...)

The use is to get a realistic value for "mask_infection_prob_decrease" and to show how "base_infection_prob" affects things.